+++ app/Http/Controllers/Admin/UserManagement.php
@@ -6,8 +6,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Controllers\Controller;
 use App\Models\User;
-use Auth;
 use jeremykenedy\LaravelRoles\Models\Role;
+use jeremykenedy\LaravelRoles\Models\Permission;
+use Auth;
 
 class UserManagement extends Controller
 {