+++ resources/assets/js/admin-lte.js
@@ -66,6 +66,8 @@ if (token) {
 // });
 
 $(document).ready(function() {
+    $('[data-toggle="tooltip"]').tooltip();
+
     $('.i-checks').iCheck({
         checkboxClass: 'icheckbox_square-blue',
         radioClass: 'iradio_square-blue'